Optionally turn off server side md5 verification
This will almost certainly improve download performance, at the cost
of delaying the diagnosis of a problem (but not the occurance) in the
unlikely even of disk corruption.  As this is used in production and
is critical it is for the moment turned OFF by default.  After an
successful experimentation phase we may change the default.

 - IF server side has config download_md5_verify = 0
 - AND client provides header X-Yars-Skip-Verify
   (any value)
   (always the case for Yars::Client)
 - THEN the server side md5 sum verification will be skipped.
 - the client is expected in this case to do a md5 sum check.

Not directly related, but also in this commit:

 - X-Use-X-Accel is renamed X-Yars-Use-X-Accel